+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
 - Deprecated ElementNotFoundException in favor of Craft's first party exception.
 - Creating an object from a record now considers any record relations populated.
 - Enhancing helper traits for easier overrides
+- ModelValidator now accepts an array of models
 
 ## 1.0.0-rc.8 - 2018-01-31
 ### Added